editoast: move rjs power restriction range to schemas
