+++ editoast/src/core/simulation.rs
@@ -1,5 +1,6 @@
 use editoast_schemas::rolling_stock::RollingStockComfortType;
 use editoast_schemas::train_schedule::Allowance;
+use editoast_schemas::train_schedule::RjsPowerRestrictionRange;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
 use utoipa::ToSchema;
@@ -8,7 +9,6 @@ use super::AsCoreRequest;
 use super::Json;
 use crate::models::train_schedule::ElectrificationRange;
 use crate::models::train_schedule::Mrsp;
-use crate::models::train_schedule::RjsPowerRestrictionRange;
 use crate::models::train_schedule::SimulationPowerRestrictionRange;
 use crate::models::train_schedule::TrainScheduleOptions;
 use crate::models::Pathfinding;
