+++ editoast/src/models/train_schedule.rs
@@ -8,6 +8,7 @@ use diesel::QueryDsl;
 use diesel_async::RunQueryDsl;
 use editoast_derive::Model;
 use editoast_schemas::train_schedule::Allowance;
+use editoast_schemas::train_schedule::RjsPowerRestrictionRange;
 use serde::Deserialize;
 use serde::Serialize;
 use utoipa::ToSchema;
@@ -25,7 +26,6 @@ use crate::DieselJson;
 
 editoast_common::schemas! {
     TrainSchedule,
-    RjsPowerRestrictionRange,
     LightTrainSchedule,
     MechanicalEnergyConsumedBaseEco,
     TrainScheduleValidation,
@@ -323,22 +323,6 @@ pub struct ResultTrain {
     pub zone_updates: Vec<ZoneUpdate>,
     pub spacing_requirements: Vec<SpacingRequirement>,
     pub routing_requirements: Vec<RoutingRequirement>,
-}
-
-/// A range along the train path where a power restriction is applied.
-#[derive(Debug, Clone, Serialize, Deserialize, ToSchema, PartialEq)]
-#[schema(example = json!({
-    "begin_position": 0.0,
-    "end_position": 1000.0,
-    "power_restriction_code": "C1US"
-}))]
-pub struct RjsPowerRestrictionRange {
-    /// Offset from the start of the path, in meters.
-    begin_position: f32,
-    /// Offset from the start of the path, in meters.
-    end_position: f32,
-    /// The power restriction code to apply.
-    power_restriction_code: String,
 }
 
 /// Options for the standalone simulation
