+++ editoast/src/views/single_simulation.rs
@@ -4,6 +4,7 @@ use actix_web::web::Json;
 use editoast_derive::EditoastError;
 use editoast_schemas::rolling_stock::RollingStockComfortType;
 use editoast_schemas::train_schedule::Allowance;
+use editoast_schemas::train_schedule::RjsPowerRestrictionRange;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
 use thiserror::Error;
@@ -20,7 +21,6 @@ use crate::error::Result;
 use crate::models::train_schedule::ElectrificationRange;
 use crate::models::train_schedule::Mrsp;
 use crate::models::train_schedule::ResultTrain;
-use crate::models::train_schedule::RjsPowerRestrictionRange;
 use crate::models::train_schedule::ScheduledPoint;
 use crate::models::train_schedule::SimulationPowerRestrictionRange;
 use crate::models::train_schedule::TrainScheduleOptions;
